split functions to lib
